Implement crt0

+++ userspace/libs/libc/include/stddef.h
@@ -5,6 +5,6 @@
 
 #pragma once
 
-typedef unsigned long long size_t;
-typedef long long ptrdiff_t;
+typedef unsigned long size_t;
+typedef long ptrdiff_t;
 #define NULL ((void *) 0)
